Add files via upload
Anomaly detection is a crucial technique in data analysis and machine learning used to identify patterns that do not conform to expected behavior within a dataset. The goal is to uncover data points, events, or observations that deviate significantly from the norm, suggesting either interesting occurrences or potential errors or outliers.